pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "archive-core"
description = "Core implementation of the Hopskotch Archive"
requires-python = ">=3.6"
dependencies = [
    "aioboto3==15.1.0",
    "aiobotocore[boto3]==2.24.0",
    "bson",
    "cython<3 ; python_version < '3.8'", # need this to work around https://github.com/fastavro/fastavro/issues/701
    "fastapi",
    "fastavro~=1.7.0 ; python_version < '3.8'", # need this to work around https://github.com/fastavro/fastavro/issues/701
    "hop-client>=0.9.0",
    "httpx",
    "psycopg[binary]",
    "psycopg_pool",
    "sqlalchemy",
    "toml",
    "uvicorn[standard]",
]
version = "0"

[tool.setuptools]
packages = ["archive"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]